[AMD] support eplb for moriep

+++ python/sglang/srt/eplb/expert_location_dispatch.py
@@ -89,7 +89,9 @@ def topk_ids_logical_to_physical(
 def _topk_ids_logical_to_physical_static(
     topk_ids: torch.Tensor, info: Optional[ExpertLocationDispatchInfo]
 ) -> torch.Tensor:
-    return info.partial_logical_to_rank_dispatch_physical_map[topk_ids]
+    return info.partial_logical_to_rank_dispatch_physical_map[topk_ids].to(
+        topk_ids.dtype
+    )
 
 
 def _topk_ids_logical_to_physical_dynamic(
